requirements.txt
# Core OCR and PDF processing
paddleocr>=2.7.0
paddlepaddle>=2.6.0
pymupdf>=1.23.0
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Web UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
langdetect>=1.0.9

# Phase 2: Content Analysis & Job Matching
# NLP and text processing
spacy>=3.7.0

# Semantic similarity (optional - for job matching)
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Development
pytest>=7.4.0

# Note: After installing, run:
# python -m spacy download en_core_web_sm
#
# For local LLM support (optional):
# Install Ollama from https://ollama.com
# Then run: ollama pull ministral:3b-instruct-2512-q4_K_M
#
# For OpenRouter API (optional):
# Set OPENROUTER_API_KEY environment variable
# Free models available: mistralai/mistral-small-3.1-24b-instruct:free
